+++ tests/test_amends.py
@@ -36,8 +36,18 @@ def test_exclude_amends_from_xml(apple_10k_edgar_search_xml_url):
 def test_include_amends_filing_retrieval(downloader, apple_filing_metadata):
     dl, download_location = downloader
 
-    # AAPL has 1 SG-13G filing and 2 SG-13G/A before 20190531
+    # AAPL has 1 SG-13G filing and 2 SG-13G/A before 19940218
     num_downloaded = dl.get_sc_13g_filings(
-        apple_filing_metadata["symbol"], 3, "20190531", True
+        apple_filing_metadata["symbol"], 3, "19940218", True
     )
     assert num_downloaded == 3
+
+
+def test_exclude_amends_filing_retrieval(downloader, apple_filing_metadata):
+    dl, download_location = downloader
+
+    # AAPL has 1 SG-13G filing and 2 SG-13G/A before 19940218
+    num_downloaded = dl.get_sc_13g_filings(
+        apple_filing_metadata["symbol"], 3, "19940218", False
+    )
+    assert num_downloaded == 1